db: fix hash lookups on 64-bit machines
The hash is 'unsigned long' everywhere which is 64-bits (on 64-bit
boxes). Fix the one instance of it being 'unsigned int' since that
would be 32-bit and cause lookup failures on 64-bit boxes.

+++ src/apk_database.h
@@ -44,7 +44,7 @@ struct apk_db_file {
 struct apk_db_dir {
 	apk_hash_node hash_node;
 
-	unsigned int hash;
+	unsigned long hash;
 	struct hlist_head files;
 	struct apk_db_dir *parent;
